requirements.txt
pydantic
python-dotenv
Django
djangorestframework
django-filter
psycopg2-binary